+++ packages/react-ai/src/components/AIConversation/context/index.ts
@@ -2,6 +2,7 @@ export { AIContextContext, AIContextProvider } from './AIContextContext';
 export { ActionsContext, ActionsProvider } from './ActionsContext';
 export { AvatarsContext, AvatarsProvider } from './AvatarsContext';
 export {
+  ConversationInputContextProps,
   ConversationInputContext,
   ConversationInput,
   ConversationInputContextProvider,
@@ -40,7 +41,11 @@ export {
   MessageRendererContext,
   useMessageRenderer,
 } from './MessageRenderContext';
-export { AttachmentProvider, AttachmentContext } from './AttachmentContext';
+export {
+  AttachmentProvider,
+  AttachmentContext,
+  AttachmentContextProps,
+} from './AttachmentContext';
 export {
   WelcomeMessageContext,
   WelcomeMessageProvider,