+++ packages/react-native-auth/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @aws-amplify/ui-react-native-auth
 
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [[`475e4a37a404bd116ec55e95579acb8f2d23c9a4`](https://github.com/aws-amplify/amplify-ui/commit/475e4a37a404bd116ec55e95579acb8f2d23c9a4)]:
+  - @aws-amplify/ui-react-native@2.3.1
+  - @aws-amplify/ui-react-core-auth@0.0.37
+
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [[`825cccc92f5c326f0f110c8439a1c2c6553ff00a`](https://github.com/aws-amplify/amplify-ui/commit/825cccc92f5c326f0f110c8439a1c2c6553ff00a), [`50cde53145bfffc97f17b6da5c650c25d992fd97`](https://github.com/aws-amplify/amplify-ui/commit/50cde53145bfffc97f17b6da5c650c25d992fd97), [`954e9be2d055a57ebddab5b902e3e6b427d09f67`](https://github.com/aws-amplify/amplify-ui/commit/954e9be2d055a57ebddab5b902e3e6b427d09f67)]:
+  - @aws-amplify/ui-react-native@2.3.0
+  - @aws-amplify/ui-react-core-auth@0.0.36
+
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [[`96be188317d72b89b6bfbc14a3d4ee76daad2a8b`](https://github.com/aws-amplify/amplify-ui/commit/96be188317d72b89b6bfbc14a3d4ee76daad2a8b)]:
+  - @aws-amplify/ui@6.7.2
+  - @aws-amplify/ui-react-core-auth@0.0.35
+  - @aws-amplify/ui-react-native@2.2.20
+
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies [[`9ab56f499f4596062fe4614a016efc3dfc88ef03`](https://github.com/aws-amplify/amplify-ui/commit/9ab56f499f4596062fe4614a016efc3dfc88ef03)]:
+  - @aws-amplify/ui@6.7.1
+  - @aws-amplify/ui-react-core-auth@0.0.34
+  - @aws-amplify/ui-react-native@2.2.19
+
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies [[`70971f310`](https://github.com/aws-amplify/amplify-ui/commit/70971f310f69717657849f29d0a9e1d993b08d9a)]:
+  - @aws-amplify/ui@6.7.0
+  - @aws-amplify/ui-react-core-auth@0.0.33
+  - @aws-amplify/ui-react-native@2.2.18
+
 ## 0.0.35
 
 ### Patch Changes