+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @aws-amplify/ui-vue
 
+## 4.2.26
+
+### Patch Changes
+
+- [#6250](https://github.com/aws-amplify/amplify-ui/pull/6250) [`27c72b51e1f11073e383eb848a99589c234abb23`](https://github.com/aws-amplify/amplify-ui/commit/27c72b51e1f11073e383eb848a99589c234abb23) Thanks [@calebpollman](https://github.com/calebpollman)! - deps: resolve all versions of nanoid to 3.3.8
+
+- Updated dependencies [[`96be188317d72b89b6bfbc14a3d4ee76daad2a8b`](https://github.com/aws-amplify/amplify-ui/commit/96be188317d72b89b6bfbc14a3d4ee76daad2a8b)]:
+  - @aws-amplify/ui@6.7.2
+
+## 4.2.25
+
+### Patch Changes
+
+- Updated dependencies [[`9ab56f499f4596062fe4614a016efc3dfc88ef03`](https://github.com/aws-amplify/amplify-ui/commit/9ab56f499f4596062fe4614a016efc3dfc88ef03)]:
+  - @aws-amplify/ui@6.7.1
+
+## 4.2.24
+
+### Patch Changes
+
+- Updated dependencies [[`70971f310`](https://github.com/aws-amplify/amplify-ui/commit/70971f310f69717657849f29d0a9e1d993b08d9a)]:
+  - @aws-amplify/ui@6.7.0
+
 ## 4.2.23
 
 ### Patch Changes